refactor tests, add GitHub actions
