+++ .github/workflows/testing.yml
@@ -1,0 +1,27 @@
+name: testing
+
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v2
+
+    - name: Set up Go
+      uses: actions/setup-go@v2
+      with:
+        go-version: 1.16
+
+    - name: Build
+      run: go build -v ./...
+
+    - name: Test
+      env: 
+        FIVETRAN_APIKEY : ${{ secrets.FIVETRAN_APIKEY }}
+        FIVETRAN_APISECRET : ${{ secrets.FIVETRAN_APISECRET }}
+      run: go test -v ./...
