+++ fivetran/resource_user_test.go
@@ -1,4 +1,4 @@
-package tests
+package fivetran_test
 
 import (
 	"context"
@@ -9,9 +9,9 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
 
-func TestFivetranUser_basic(t *testing.T) {
+func TestResourceUserE2E(t *testing.T) {
 	resource.Test(t, resource.TestCase{
-		PreCheck: func() {},
+		PreCheck:     func() {},
 		Providers:    testProviders,
 		CheckDestroy: testFivetranUserResourceDestroy,
 		Steps: []resource.TestStep{
@@ -19,8 +19,8 @@ func TestFivetranUser_basic(t *testing.T) {
 				Config: `
 		   	resource "fivetran_user" "userjohn" {
 				 provider = fivetran-provider
-			     email = "john.black@testmail.com"
-			     family_name = "Black"
+			     email = "john.fox@testmail.com"
+			     family_name = "Fox"
 			     given_name = "John"
 			     phone = "+19876543210"
 			     picture = "https://myPicturecom"
@@ -28,8 +28,8 @@ func TestFivetranUser_basic(t *testing.T) {
 		  `,
 				Check: resource.ComposeAggregateTestCheckFunc(
 					testFivetranUserResourceCreate(t, "fivetran_user.userjohn"),
-					resource.TestCheckResourceAttr("fivetran_user.userjohn", "email", "john.black@testmail.com"),
-					resource.TestCheckResourceAttr("fivetran_user.userjohn", "family_name", "Black"),
+					resource.TestCheckResourceAttr("fivetran_user.userjohn", "email", "john.fox@testmail.com"),
+					resource.TestCheckResourceAttr("fivetran_user.userjohn", "family_name", "Fox"),
 					resource.TestCheckResourceAttr("fivetran_user.userjohn", "given_name", "John"),
 					resource.TestCheckResourceAttr("fivetran_user.userjohn", "phone", "+19876543210"),
 					resource.TestCheckResourceAttr("fivetran_user.userjohn", "picture", "https://myPicturecom"),
@@ -39,8 +39,8 @@ func TestFivetranUser_basic(t *testing.T) {
 				Config: `
 		   	resource "fivetran_user" "userjohn" {
 				provider = fivetran-provider
-				email = "john.black@testmail.com"
-				family_name = "White"
+				email = "john.fox@testmail.com"
+				family_name = "Connor"
 				given_name = "Jane"
 				phone = "+19876543219"
 				picture = "https://yourPicturecom"
@@ -48,8 +48,8 @@ func TestFivetranUser_basic(t *testing.T) {
 		  `,
 				Check: resource.ComposeAggregateTestCheckFunc(
 					testFivetranUserResourceUpdate(t, "fivetran_user.userjohn"),
-					resource.TestCheckResourceAttr("fivetran_user.userjohn", "email", "john.black@testmail.com"),
-					resource.TestCheckResourceAttr("fivetran_user.userjohn", "family_name", "White"),
+					resource.TestCheckResourceAttr("fivetran_user.userjohn", "email", "john.fox@testmail.com"),
+					resource.TestCheckResourceAttr("fivetran_user.userjohn", "family_name", "Connor"),
 					resource.TestCheckResourceAttr("fivetran_user.userjohn", "given_name", "Jane"),
 					resource.TestCheckResourceAttr("fivetran_user.userjohn", "phone", "+19876543219"),
 					resource.TestCheckResourceAttr("fivetran_user.userjohn", "picture", "https://yourPicturecom"),
